uyeah2

+++ lib/audio_output/src/I2SOutput.cpp
@@ -1,12 +1,10 @@
 
 #include "I2SOutput.h"
 
-I2SOutput::I2SOutput(i2s_port_t i2s_port, i2s_pin_config_t &i2s_pins) : Output(i2s_port), m_i2s_pins(i2s_pins)
-{
-}
+I2SOutput::I2SOutput(i2s_port_t i2s_port, i2s_pin_config_t &i2s_pins)
+    : Output(i2s_port), m_i2s_pins(i2s_pins) {}
 
-void I2SOutput::start(uint32_t sample_rate)
-{
+void I2SOutput::start(uint16_t sample_rate) {
     // i2s config for writing both channels of I2S
     i2s_config_t i2s_config = {
         .mode = (i2s_mode_t)(I2S_MODE_MASTER | I2S_MODE_TX),
@@ -29,8 +27,8 @@ void I2SOutput::start(uint32_t sample_rate)
         .tx_desc_auto_clear = true,
         .fixed_mclk = 0,
 #if ESP_IDF_VERSION >= ESP_IDF_VERSION_VAL(4, 4, 1)
-        .mclk_multiple = I2S_MCLK_MULTIPLE_DEFAULT, // Unused
-        .bits_per_chan = I2S_BITS_PER_CHAN_DEFAULT  // Use bits per sample
+        .mclk_multiple = I2S_MCLK_MULTIPLE_DEFAULT,  // Unused
+        .bits_per_chan = I2S_BITS_PER_CHAN_DEFAULT   // Use bits per sample
 #endif
     };
     // install and start i2s driver
